+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Event group creation view
 - Event group editing view
 - View for adding an event to an event group
+- Breadcrumb to occurrence detail view
 
 ### Changed